requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
Django==3.1.5
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
django-rest-knox==4.1.0
django-social-share==2.2.1
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
markdown2==2.3.10
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.1.0
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
